% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pclake.R
\name{run_pclake}
\alias{run_pclake}
\title{Run the PCLake model}
\usage{
run_pclake(sim_folder = ".", par_file = "langtjern.par", verbose = TRUE)
}
\arguments{
\item{sim_folder}{the directory where simulation files are contained}

\item{par_file}{the parameter file that needs to be run}

\item{verbose}{Save output as character vector. Defaults to FALSE}
}
\description{
This runs the PCLake model on the specific simulation stored in \code{sim_folder}.
The specified \code{sim_folder} must contain a valid .par file.
}
\examples{
sim_folder <- system.file('extdata', package = 'SimstratR')
run_simstrat(sim_folder, par_file = 'langtjern.par')
}
\author{
Jorrit Mesman, Tadhg Moore
}
\keyword{methods}
